+++ pgql_a_example.py
@@ -34,15 +34,14 @@ def handle_result(result: SuiRpcResult) -> SuiRpcResult:
 async def do_coin_meta(client: AsyncSuiGQLClient):
     """Fetch meta data about coins, includes supply."""
     # Defaults to 0x2::sui::SUI
-    handle_result(await client.execute_query(with_query_node=qn.GetCoinMetaData()))
+    handle_result(await client.execute_query_node(with_node=qn.GetCoinMetaData()))
 
 
 async def do_coins_for_type(client: AsyncSuiGQLClient):
     """Fetch coins of specific type for owner."""
     handle_result(
-        await client.execute_query(
-            # GetAllCoinsOfType requires a coin type
-            with_query_node=qn.GetCoins(
+        await client.execute_query_node(
+            with_node=qn.GetCoins(
                 owner=client.config.active_address.address,
                 coin_type="0x2::sui::SUI",
             )
@@ -53,9 +52,8 @@ async def do_coins_for_type(client: AsyncSuiGQLClient):
 async def do_gas(client: AsyncSuiGQLClient):
     """Fetch 0x2::sui::SUI (default) for owner."""
     result = handle_result(
-        await client.execute_query(
-            # GetAllCoins defaults to "0x2::sui::SUI"
-            with_query_node=qn.GetCoins(owner=client.config.active_address.address)
+        await client.execute_query_node(
+            with_node=qn.GetCoins(owner=client.config.active_address.address)
         )
     )
     if result.is_ok():
@@ -67,7 +65,7 @@ async def do_gas(client: AsyncSuiGQLClient):
 async def do_sysstate(client: AsyncSuiGQLClient):
     """Fetch the most current system state summary."""
     handle_result(
-        await client.execute_query(with_query_node=qn.GetLatestSuiSystemState())
+        await client.execute_query_node(with_node=qn.GetLatestSuiSystemState())
     )
 
 
@@ -76,16 +74,14 @@ async def do_all_balances(client: AsyncSuiGQLClient):
 
     Demonstrates paging as well
     """
-    result = await client.execute_query(
-        with_query_node=qn.GetAllCoinBalances(
-            owner=client.config.active_address.address
-        )
+    result = await client.execute_query_node(
+        with_node=qn.GetAllCoinBalances(owner=client.config.active_address.address)
     )
     handle_result(result)
     if result.is_ok():
         while result.result_data.next_cursor.hasNextPage:
-            result = await client.execute_query(
-                with_query_node=qn.GetAllCoinBalances(
+            result = await client.execute_query_node(
+                with_node=qn.GetAllCoinBalances(
                     owner=client.config.active_address.address,
                     next_page=result.next_cursor,
                 )
@@ -97,15 +93,15 @@ async def do_all_balances(client: AsyncSuiGQLClient):
 async def do_object(client: AsyncSuiGQLClient):
     """Fetch specific object data."""
     handle_result(
-        await client.execute_query(with_query_node=qn.GetObject(object_id="0x6"))
+        await client.execute_query_node(with_node=qn.GetObject(object_id="0x6"))
     )
 
 
 async def do_objects(client: AsyncSuiGQLClient):
     """Fetch all objects help by owner."""
     handle_result(
-        await client.execute_query(
-            with_query_node=qn.GetObjectsOwnedByAddress(
+        await client.execute_query_node(
+            with_node=qn.GetObjectsOwnedByAddress(
                 owner=client.config.active_address.address
             )
         )
@@ -117,8 +113,8 @@ async def do_past_object(client: AsyncSuiGQLClient):
     To run, change the objectID str and version int.
     """
     handle_result(
-        await client.execute_query(
-            with_query_node=qn.GetPastObject(
+        await client.execute_query_node(
+            with_node=qn.GetPastObject(
                 object_id="0xdfa764b29d303acecc801828839108ea81a45e93c3b9ccbe05b0d9a697a2a9ed",
                 version=17078252,
             )
@@ -137,8 +133,8 @@ async def do_multiple_past_object(client: AsyncSuiGQLClient):
         }
     ]
     handle_result(
-        await client.execute_query(
-            with_query_node=qn.GetMultiplePastObjects(for_versions=past_objects)
+        await client.execute_query_node(
+            with_node=qn.GetMultiplePastObjects(for_versions=past_objects)
         )
     )
 
@@ -149,8 +145,8 @@ async def do_objects_for(client: AsyncSuiGQLClient):
     These are test IDs, replace to run.
     """
     handle_result(
-        await client.execute_query(
-            with_query_node=qn.GetMultipleObjects(
+        await client.execute_query_node(
+            with_node=qn.GetMultipleObjects(
                 object_ids=[
                     "0x0847e1e02965e3f6a8b237152877a829755fd2f7cfb7da5a859f203a8d4316f0",
                     "0x68e961e3af906b160e1ff21137304537fa6b31f5a4591ef3acf9664eb6e3cd2b",
@@ -167,8 +163,8 @@ async def do_dynamics(client: AsyncSuiGQLClient):
     This is test ID, replace to run.
     """
     handle_result(
-        await client.execute_query(
-            with_query_node=qn.GetDynamicFields(
+        await client.execute_query_node(
+            with_node=qn.GetDynamicFields(
                 object_id="0xdfa764b29d303acecc801828839108ea81a45e93c3b9ccbe05b0d9a697a2a9ed"
             )
         )
@@ -178,8 +174,8 @@ async def do_dynamics(client: AsyncSuiGQLClient):
 async def do_event(client: AsyncSuiGQLClient):
     """."""
     handle_result(
-        await client.execute_query(
-            with_query_node=qn.GetEvents(event_filter={"sender": "0x0"})
+        await client.execute_query_node(
+            with_node=qn.GetEvents(event_filter={"sender": "0x0"})
         )
     )
 
@@ -200,10 +196,8 @@ async def do_chain_id(client: AsyncSuiGQLClient):
 async def do_tx(client: AsyncSuiGQLClient):
     """Fetch specific transaction by it's digest."""
     handle_result(
-        await client.execute_query(
-            with_query_node=qn.GetTx(
-                digest="A8kCT1n8dmCWchz5WnKPsQ8x7U49ExgMEWJ13nRULpiz"
-            )
+        await client.execute_query_node(
+            with_node=qn.GetTx(digest="A8kCT1n8dmCWchz5WnKPsQ8x7U49ExgMEWJ13nRULpiz")
         )
     )
 
@@ -213,7 +207,7 @@ async def do_txs(client: AsyncSuiGQLClient):
 
     We loop through 3 pages.
     """
-    result = await client.execute_query(with_query_node=qn.GetMultipleTx())
+    result = await client.execute_query_node(with_node=qn.GetMultipleTx())
     handle_result(result)
     if result.is_ok():
         max_page = 3
@@ -221,10 +215,8 @@ async def do_txs(client: AsyncSuiGQLClient):
         while True:
             in_page += 1
             if in_page < max_page and result.result_data.next_cursor:
-                result = await client.execute_query(
-                    with_query_node=qn.GetMultipleTx(
-                        next_page=result.result_data.next_cursor
-                    )
+                result = await client.execute_query_node(
+                    with_node=qn.GetMultipleTx(next_page=result.result_data.next_cursor)
                 )
                 handle_result(result)
             else:
@@ -236,7 +228,7 @@ async def do_staked_sui(client: AsyncSuiGQLClient):
     """."""
     owner = client.config.active_address.address
     handle_result(
-        await client.execute_query(with_query_node=qn.GetDelegatedStakes(owner=owner))
+        await client.execute_query_node(with_node=qn.GetDelegatedStakes(owner=owner))
     )
 
 
@@ -244,21 +236,17 @@ async def do_latest_cp(client: AsyncSuiGQLClient):
     """."""
     qnode = qn.GetLatestCheckpointSequence()
     # print(qnode.query_as_string())
-    handle_result(await client.execute_query(with_query_node=qnode))
+    handle_result(await client.execute_query_node(with_node=qnode))
 
 
 async def do_sequence_cp(client: AsyncSuiGQLClient):
     """."""
-    result = await client.execute_query(
-        with_query_node=qn.GetLatestCheckpointSequence()
-    )
+    result = await client.execute_query_node(with_node=qn.GetLatestCheckpointSequence())
     if result.is_ok():
         cp: ptypes.CheckpointGQL = result.result_data
         handle_result(
-            await client.execute_query(
-                with_query_node=qn.GetCheckpointBySequence(
-                    sequence_number=cp.sequence_number
-                )
+            await client.execute_query_node(
+                with_node=qn.GetCheckpointBySequence(sequence_number=cp.sequence_number)
             )
         )
     else:
@@ -267,14 +255,12 @@ async def do_sequence_cp(client: AsyncSuiGQLClient):
 
 async def do_digest_cp(client: AsyncSuiGQLClient):
     """."""
-    result = await client.execute_query(
-        with_query_node=qn.GetLatestCheckpointSequence()
-    )
+    result = await client.execute_query_node(with_node=qn.GetLatestCheckpointSequence())
     if result.is_ok():
         cp: ptypes.CheckpointGQL = result.result_data
         handle_result(
-            await client.execute_query(
-                with_query_node=qn.GetCheckpointByDigest(digest=cp.digest)
+            await client.execute_query_node(
+                with_node=qn.GetCheckpointByDigest(digest=cp.digest)
             )
         )
     else:
@@ -283,19 +269,19 @@ async def do_digest_cp(client: AsyncSuiGQLClient):
 
 async def do_checkpoints(client: AsyncSuiGQLClient):
     """."""
-    handle_result(await client.execute_query(with_query_node=qn.GetCheckpoints()))
+    handle_result(await client.execute_query_node(with_node=qn.GetCheckpoints()))
 
 
 async def do_refgas(client: AsyncSuiGQLClient):
     """Fetch the most current system state summary."""
-    handle_result(await client.execute_query(with_query_node=qn.GetReferenceGasPrice()))
+    handle_result(await client.execute_query_node(with_node=qn.GetReferenceGasPrice()))
 
 
 async def do_nameservice(client: AsyncSuiGQLClient):
     """Fetch the most current system state summary."""
     handle_result(
-        await client.execute_query(
-            with_query_node=qn.GetNameServiceAddress(name="gql-frank")
+        await client.execute_query_node(
+            with_node=qn.GetNameServiceAddress(name="gql-frank")
         )
     )
 
@@ -303,28 +289,26 @@ async def do_nameservice(client: AsyncSuiGQLClient):
 async def do_owned_nameservice(client: AsyncSuiGQLClient):
     """Fetch the most current system state summary."""
     handle_result(
-        await client.execute_query(
-            with_query_node=qn.GetNameServiceNames(
-                owner=client.config.active_address.address
-            )
+        await client.execute_query_node(
+            with_node=qn.GetNameServiceNames(owner=client.config.active_address.address)
         )
     )
 
 
 async def do_validators_apy(client: AsyncSuiGQLClient):
     """Fetch the most current validators apy and identity."""
-    handle_result(await client.execute_query(with_query_node=qn.GetValidatorsApy()))
+    handle_result(await client.execute_query_node(with_node=qn.GetValidatorsApy()))
 
 
 async def do_validators(client: AsyncSuiGQLClient):
     """Fetch the most current validator detail."""
-    handle_result(await client.execute_query(with_query_node=qn.GetCurrentValidators()))
+    handle_result(await client.execute_query_node(with_node=qn.GetCurrentValidators()))
 
 
 async def do_protcfg(client: AsyncSuiGQLClient):
     """Fetch the most current system state summary."""
     handle_result(
-        await client.execute_query(with_query_node=qn.GetProtocolConfig(version=30))
+        await client.execute_query_node(with_node=qn.GetProtocolConfig(version=30))
     )
 
 
@@ -333,8 +317,8 @@ async def do_struct(client: AsyncSuiGQLClient):
 
     This is a testnet object!!!
     """
-    result = await client.execute_query(
-        with_query_node=qn.GetStructure(
+    result = await client.execute_query_node(
+        with_node=qn.GetStructure(
             package="0x2",
             module_name="coin",
             structure_name="CoinMetadata",
@@ -349,8 +333,8 @@ async def do_structs(client: AsyncSuiGQLClient):
 
     This is a testnet object!!!
     """
-    result = await client.execute_query(
-        with_query_node=qn.GetStructures(
+    result = await client.execute_query_node(
+        with_node=qn.GetStructures(
             package="0x2",
             module_name="coin",
         )
@@ -364,8 +348,8 @@ async def do_func(client: AsyncSuiGQLClient):
 
     This is a testnet object!!!
     """
-    result = await client.execute_query(
-        with_query_node=qn.GetFunction(
+    result = await client.execute_query_node(
+        with_node=qn.GetFunction(
             package="0x2",
             module_name="coin",
             function_name="join",
@@ -380,8 +364,8 @@ async def do_funcs(client: AsyncSuiGQLClient):
 
     This is a testnet object!!!
     """
-    result = await client.execute_query(
-        with_query_node=qn.GetFunctions(
+    result = await client.execute_query_node(
+        with_node=qn.GetFunctions(
             package="0x2",
             module_name="coin",
         )
@@ -395,8 +379,8 @@ async def do_module(client: AsyncSuiGQLClient):
 
     This is a testnet object!!!
     """
-    result = await client.execute_query(
-        with_query_node=qn.GetModule(
+    result = await client.execute_query_node(
+        with_node=qn.GetModule(
             package="0x2",
             module_name="coin",
         )
@@ -410,8 +394,8 @@ async def do_package(client: AsyncSuiGQLClient):
 
     This is a testnet object!!!
     """
-    result = await client.execute_query(
-        with_query_node=qn.GetPackage(
+    result = await client.execute_query_node(
+        with_node=qn.GetPackage(
             package="0x2",
         )
     )
@@ -431,8 +415,8 @@ async def do_dry_run(client: AsyncSuiGQLClient):
         tx_data = await txer.get_transaction_data()
         tx_b64 = base64.b64encode(tx_data.serialize()).decode()
         handle_result(
-            await client.execute_query(
-                with_query_node=qn.DryRunTransaction(tx_bytestr=tx_b64)
+            await client.execute_query_node(
+                with_node=qn.DryRunTransaction(tx_bytestr=tx_b64)
             )
         )
 
@@ -452,10 +436,8 @@ async def do_execute(client: AsyncSuiGQLClient):
         rsig_array = [x.value for x in sig_array.array]
         print(rsig_array)
         handle_result(
-            await client.execute_query(
-                with_query_node=qn.ExecuteTransaction(
-                    tx_bytestr=tx_b64, sig_array=rsig_array
-                )
+            await client.execute_query_node(
+                with_node=qn.ExecuteTransaction(tx_bytestr=tx_b64, sig_array=rsig_array)
             )
         )